+++ src/main/java/com/jokim/sivillage/api/purchase/application/PurchaseService.java
@@ -1,12 +1,12 @@
 package com.jokim.sivillage.api.purchase.application;
 
-import com.jokim.sivillage.api.purchase.dto.PurchaseRequestDto;
-import com.jokim.sivillage.api.purchase.dto.PurchaseResponseDto;
+import com.jokim.sivillage.api.purchase.dto.in.PurchaseProductRequestDto;
+import com.jokim.sivillage.api.purchase.dto.out.PurchaseResponseDto;
 import java.util.List;
 
 public interface PurchaseService {
 
-    void purchaseProduct(PurchaseRequestDto purchaseRequestDto);
+    void purchaseProduct(PurchaseProductRequestDto purchaseProductRequestDto);
 
     List<PurchaseResponseDto> getPurchaseSheet(String accessToken);
